source: refactor source into separate package
Problem: the source APIs are rich and useful but currently tangled with k8s
logic.
Solution: to extend the node-feature-discovery APIs to be used by more
projects, and not require a huge dependency (Kubernetes upstream) it
is relatively simple to refactor the structure to maintain them in
a separate module. This change includes the following modifications:
- use of git-filter-repo to maintain git history for source and pkg
- update of types.go to only include interfaces necessary for source
- removal of kubebuilder (or similar) logic and grpc and kubeconfig, etc.
- entire removal of nodefeaturerule and Rule/Match/custom source
- flags/memory_resources|grpc|kubeconfig|metrics|tls not relevant here
- klog is replaced with slog, which is available with golang 1.21
Complete changes are included in README.md with this commit.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>